lab2

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,13 +3,14 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, int bottom) {
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
 
     public static void main(String[] args) {
+        /**
         int t = 10;
         int b = 2;
         int result = divideThenRound(t, b);
@@ -19,8 +20,8 @@ public class DebugExercise1 {
         int b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
-
-        int t3 = 3;
+        */
+        float t3 = 3;
         int b3 = 4;
         int result3 = divideThenRound(t3, b3);
         System.out.println("round(" + t3 + "/" + b3 + ")=" + result3);
